solved some issues and deleted unecessary files
